chore: Fix tests and linter
