+++ libs/backend/db-utils/src/transform/dataObject.ts
@@ -1,8 +1,9 @@
-import { ExcludeFalsy, OmitFunctions } from '@scrapper-gate/shared/common';
+import { ExcludeFalsy } from '@scrapper-gate/shared/common';
 import {
   DataObjectConstructor,
   DataObject,
 } from '@scrapper-gate/shared/data-object';
+import { OmitFunctions } from '@scrapper-gate/shared/data-structures';
 import { PartialDeep } from 'type-fest';
 import { ValueTransformer } from 'typeorm';
