+++ libs/shared/common/src/types.ts
@@ -5,25 +5,6 @@ export interface QueryResult<T> {
   items: T[];
 }
 
-export type Dictionary = Record<string, unknown>;
-
-export type OmitFunctions<T> = Pick<
-  T,
-  { [K in keyof T]: T[K] extends (_: unknown) => unknown ? never : K }[keyof T]
->;
-
-export interface Jsonable {
-  toJSON(): Dictionary;
-}
-
-export interface Clonable {
-  clone: () => this;
-}
-
-export interface WithValue<T> {
-  valueOf: () => T;
-}
-
 export type Exists<T> = Exclude<T, null | undefined>;
 
 export type ExistsInObject<T> = {
