+++ libs/shared/data-structures/README.md
@@ -1,0 +1,7 @@
+# shared-data-structures
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-data-structures` to execute the unit tests via [Jest](https://jestjs.io).
