+++ tsconfig.base.json
@@ -126,6 +126,9 @@
       "@scrapper-gate/shared/data-object": [
         "libs/shared/data-object/src/index.ts"
       ],
+      "@scrapper-gate/shared/data-structures": [
+        "libs/shared/data-structures/src/index.ts"
+      ],
       "@scrapper-gate/shared/domain/auth": [
         "libs/shared/domain/auth/src/index.ts"
       ],
